Demonstrate git commit signatures
This git commit is signed by my PGP key. You should be able to see the
signature on it using the --show-signature option to git log. (you may
need to run gpg --recv-key 7FAB114267E4FA04 first to get my PGP key from
the keyservers)

This should be the same commit ID as the tag v0.1.1 - beyond that I
haven't done any special verification or auditing of the code.